pyproject.toml
[project]
name = "lovelaice"
version = "0.7.0"
description = "A local-first coding agent for the terminal, with Python-defined commands and tools."
readme = "README.md"
authors = [
    { name = "Alejandro Piad", email = "apiad@apiad.net" }
]
requires-python = ">=3.13"
dependencies = [
    "lingo-ai==1.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
lovelaice="lovelaice.cli:app"
